+++ lib/rgeo/geographic/projector.rb
@@ -8,7 +8,7 @@
 
 module RGeo
   module Geographic
-    class Proj4Projector # :nodoc:
+    class Projector # :nodoc:
       def initialize(geography_factory, projection_factory)
         @geography_factory = geography_factory
         @projection_factory = projection_factory
@@ -42,11 +42,11 @@ module RGeo
           new(geography_factory, projection_factory)
         end
 
-        def create_from_proj4(geography_factory, proj4, opts = {})
+        def create_from_opts(geography_factory, opts = {})
           projection_factory =
             Cartesian.preferred_factory(
-              proj4: proj4,
               coord_sys: opts[:coord_sys], srid: opts[:srid],
+              coord_sys_class: opts[:coord_sys_class],
               buffer_resolution: opts[:buffer_resolution],
               has_z_coordinate: opts[:has_z_coordinate],
               has_m_coordinate: opts[:has_m_coordinate],